requirements.txt
Django>=4.2,<5.3
Pillow>=10.3
django-storages[azure]>=1.14
